# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NoteLingua API",
    "version": "1.0.0",
    "description": "API documentation for NoteLingua - A vocabulary learning platform",
    "contact": {
      "name": "NoteLingua Team",
      "email": "support@notelingua.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.notelingua.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token in format: Bearer <token>"
      }
    },
    "schemas": {
...
      "Document": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "Reference to user ID",
            "required": true
          },
          "file_hash": {
            "type": "string",
            "required": true
          },
          "file_name": {
            "type": "string",
            "maxLength": [
              255,
              "File name cannot exceed 255 characters"
            ],
            "required": true
          },
          "notes": {
            "type": "object",
            "properties": {
              "word": {
                "type": "string",
                "maxLength": [
                  100,
                  "Word cannot exceed 100 characters"
                ],
                "required": true
              },
              "position.page": {
                "type": "number",
                "minimum": 1
              },
              "position.start": {
                "type": "number",
                "minimum": 0
              },
              "position.end": {
                "type": "number",
                "minimum": 0
              },
              "meaning": {
                "type": "string",
                "maxLength": [
                  1000,
                  "Meaning cannot exceed 1000 characters"
                ],
                "required": true
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "source_tag": {
                "type": "string",
                "maxLength": [
                  50,
                  "Source tag cannot exceed 50 characters"
                ]
              }
            },
            "required": [
              "word",
              "meaning"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "created timestamp"
          },
          "id": {
            "type": "string",
            "description": "Reference to  ID",
            "example": "507f1f77bcf86cd799439011"
          }
        },
        "required": [
          "user_id",
          "file_hash",
          "file_name"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type",
            "example": "Validation Error"
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Please check your input data"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "Please provide a valid email"
                }
              }
            },
            "description": "Validation error details"
          }
        }
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Array of data items"
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "description": "Current page number",
                "example": 1
              },
              "limit": {
                "type": "integer",
                "description": "Items per page",
                "example": 10
              },
              "total": {
                "type": "integer",
                "description": "Total items count",
                "example": 100
              },
              "pages": {
                "type": "integer",
                "description": "Total pages count",
                "example": 10
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/documents/import": {
      "post": {
        "summary": "Upload and import document",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "document"
                ],
                "properties": {
                  "document": {
                    "type": "string",
                    "format": "binary",
                    "description": "Document file to upload"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document already exists for this user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Document already exists for this user"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Document imported successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Document imported successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "file_hash": {
                          "type": "string",
                          "example": "abc123def456"
                        },
                        "file_name": {
                          "type": "string",
                          "example": "document.pdf"
                        },
                        "file_path": {
                          "type": "string",
                          "example": "uploads/documents/document.pdf"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No file uploaded or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/documents": {
      "get": {
        "summary": "Get user's documents with pagination",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Number of documents per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search in document names"
          }
        ],
        "responses": {
          "200": {
            "description": "Documents retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Document"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}": {
      "get": {
        "summary": "Get document by ID with notes",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Document retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Document"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete document and its associated notes",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Document deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Document and associated notes deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}